Revert "Synchronise hour and minute hands of RadialTimePicker"
This reverts commit 16c69f15fba676ba3450d9eeedb314db3bf8b82c.

+++ Code/RadialControls/Utilities/Conversion/HoursDegreesConverter.cs
@@ -7,7 +7,7 @@ namespace RadialControls.Utilities
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            return (((double) value) / 12) * 360;
+            return (((double) (int) value) / 12) * 360;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
